moved availability to topics from prompts, seperated manage to sections
